load worker in main and pass paths so it works both on compiled and non-compiled
