+++ index.html
@@ -19,9 +19,12 @@
 		<script id="main" type="module">
 			import {main} from "./src/main.js";
 			main(document.body, {
-                wasm: "lib/olm/olm.wasm",
-                legacyBundle: "lib/olm/olm_legacy.js",
-                wasmBundle: "lib/olm/olm.js",
+                worker: "src/worker.js",
+                olm: {
+                    wasm: "lib/olm/olm.wasm",
+                    legacyBundle: "lib/olm/olm_legacy.js",
+                    wasmBundle: "lib/olm/olm.js",
+                }
             });
 		</script>
         <script id="service-worker" type="disabled">
